requirements.txt
# SOL/USDT Paper Trader — Python dependencies
# Install with:  pip install -r requirements.txt

# Core
numpy>=1.24
pandas>=2.0
requests>=2.31

# Google Sheets (optional — bot runs without it)
gspread>=6.0
google-auth>=2.28
